cleaning part1
