+++ pkg/export/go.mod
@@ -2,21 +2,4 @@ module export
 
 go 1.22.5
 
-require (
-	github.com/boombuler/barcode v1.0.1 // indirect
-	github.com/f-amaral/go-async v0.3.0 // indirect
-	github.com/google/uuid v1.5.0 // indirect
-	github.com/hhrutter/lzw v1.0.0 // indirect
-	github.com/hhrutter/tiff v1.0.1 // indirect
-	github.com/johnfercher/go-tree v1.0.5 // indirect
-	github.com/johnfercher/maroto v1.0.0 // indirect
-	github.com/johnfercher/maroto/v2 v2.1.2 // indirect
-	github.com/jung-kurt/gofpdf v1.16.2 // indirect
-	github.com/mattn/go-runewidth v0.0.15 // indirect
-	github.com/pdfcpu/pdfcpu v0.6.0 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	github.com/rivo/uniseg v0.4.4 // indirect
-	golang.org/x/image v0.18.0 // indirect
-	golang.org/x/text v0.16.0 // indirect
-	gopkg.in/yaml.v2 v2.4.0 // indirect
-)
+require github.com/jung-kurt/gofpdf v1.16.2
